+++ src/aks-preview/azext_aks_preview/tests/latest/test_aks_commands.py
@@ -3141,7 +3141,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             checks=[self.is_empty()],
         ) 
 
-    @live_only()
     @AllowLargeResponse()
     @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="")
     def test_aks_nodepool_add_with_localdns_config(self, resource_group, resource_group_location):
@@ -3188,7 +3187,6 @@ class AzureKubernetesServiceScenarioTest(ScenarioTest):
             checks=[self.is_empty()],
         )
 
-    @live_only()
     @AllowLargeResponse()
     @AKSCustomResourceGroupPreparer(random_name_length=17, name_prefix="clitest", location="westus2")
     def test_aks_nodepool_update_with_localdns_config(self, resource_group, resource_group_location):