+++ src/aks-preview/HISTORY.rst
@@ -10,14 +10,13 @@ If there is no rush to release a new version, please just add a description of t
 To release a new version, please select a new version number (usually plus 1 to last patch version, X.Y.Z -> Major.Minor.Patch, more details in `\doc <https://semver.org/>`_), and then add a new section named as the new version number in this file, the content should include the new modifications and everything from the *Pending* section. Finally, update the `VERSION` variable in `setup.py` with this new version number.
 
 Pending
-
 +++++++
+* Add support for `ManagedSystem` Agent Pool Mode.
+
 18.0.0b19
 +++++++
 - Added  `--localdns-config` to `az aks nodepool add`` and  to `az aks nodepool update` to support configuring a local DNS profile for agentpools.
 
-+++++++
-* Add support for `ManagedSystem` Agent Pool Mode.
 
 18.0.0b18
 +++++++